+++ crates/error/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Minor
 
+- Implement `core::error::Error` for `Error`
 - Add `Error::new_const()` and make `Error::{space,code}()` const
-- Add `std` feature to implement `std::error::Error`
 - Add an `OutOfBounds` error
 
 ### Patch